637 : Moved configurations to settings end point
